Change to all use Scanner
