+++ java/core/src/main/java/com/lancedb/lance/ipc/ScanOptions.java
@@ -23,6 +23,7 @@ import org.apache.arrow.util.Preconditions;
  * Lance scan options.
  */
 public class ScanOptions {
+  private final Optional<Integer> fragmentId;
   private final Optional<Long> batchSize;
   private final Optional<List<String>> columns;
   private final Optional<String> filter;
@@ -31,6 +32,7 @@ public class ScanOptions {
   /**
    * Constructor for LanceScanOptions.
    *
+   * @param fragmentId the fragment to scan
    * @param batchSize Maximum row number of each returned ArrowRecordBatch.
    *                  Optional, use Optional.empty() if unspecified.
    * @param columns   (Optional) Projected columns. Optional.empty() for scanning all columns.
@@ -38,14 +40,25 @@ public class ScanOptions {
    * @param filter    (Optional) Filter expression. Optional.empty() for no filter.
    * @param substraitFilter (Optional) Substrait filter expression.
    */
-  public ScanOptions(Optional<Long> batchSize, Optional<List<String>> columns,
-      Optional<String> filter, Optional<ByteBuffer> substraitFilter) {
+  public ScanOptions(Optional<Integer> fragmentId, Optional<Long> batchSize,
+      Optional<List<String>> columns, Optional<String> filter,
+      Optional<ByteBuffer> substraitFilter) {
     Preconditions.checkArgument(!(filter.isPresent() && substraitFilter.isPresent()),
         "cannot set both substrait filter and string filter");
+    this.fragmentId = fragmentId;
     this.batchSize = batchSize;
     this.columns = columns;
     this.filter = filter;
     this.substraitFilter = substraitFilter;
+  }
+
+  /**
+   * Get the fragment id.
+   *
+   * @return Optional containing the fragment id if specified, otherwise empty.
+   */
+  public Optional<Integer> getFragmentId() {
+    return fragmentId;
   }
 
   /**
@@ -88,10 +101,37 @@ public class ScanOptions {
    * Builder for constructing LanceScanOptions.
    */
   public static class Builder {
+    private Optional<Integer> fragmentId = Optional.empty();
     private Optional<Long> batchSize = Optional.empty();
     private Optional<List<String>> columns = Optional.empty();
     private Optional<String> filter = Optional.empty();
     private Optional<ByteBuffer> substraitFilter = Optional.empty();
+
+    public Builder() {}
+
+    /**
+     * Create a builder from another scan options.
+     *
+     * @param options another scan options
+     */
+    public Builder(ScanOptions options) {
+      this.fragmentId = options.getFragmentId();
+      this.batchSize = options.getBatchSize();
+      this.columns = options.getColumns();
+      this.filter = options.getFilter();
+      this.substraitFilter = options.getSubstraitFilter();
+    }
+
+    /**
+     * Set the fragment ID.
+     *
+     * @param fragmentId fragment id to scan
+     * @return Builder instance for method chaining.
+     */
+    public Builder fragmentId(int fragmentId) {
+      this.fragmentId = Optional.of(fragmentId);
+      return this;
+    }
 
     /**
      * Set the batch size.
@@ -146,7 +186,7 @@ public class ScanOptions {
      * @return LanceScanOptions instance with the specified parameters.
      */
     public ScanOptions build() {
-      return new ScanOptions(batchSize, columns, filter, substraitFilter);
+      return new ScanOptions(fragmentId, batchSize, columns, filter, substraitFilter);
     }
   }
 }
